+++ src/test/java/ch/newinstance/plugin/mavendependencychecker/util/VersionComparatorTest.java
@@ -1,5 +1,6 @@
 package ch.newinstance.plugin.mavendependencychecker.util;
 
+import ch.newinstance.plugin.mavendependencychecker.config.MavenDependencyCheckerSettings;
 import ch.newinstance.plugin.mavendependencychecker.model.DependencyUpdateResult;
 import org.apache.maven.model.Dependency;
 import org.apache.maven.model.Plugin;
@@ -8,6 +9,10 @@ import org.jetbrains.idea.maven.model.MavenPlugin;
 import org.json.JSONArray;
 import org.json.JSONObject;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
 
 import java.util.Collections;
 import java.util.List;
@@ -15,14 +20,23 @@ import java.util.List;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.mockito.Mockito.when;
 
+@ExtendWith(MockitoExtension.class)
 class VersionComparatorTest {
 
+    @Mock
+    private MavenDependencyCheckerSettings settings;
+
+    @InjectMocks
     private VersionComparator testee;
+
+    // Dependencies //
 
     @Test
     void compareDependencyVersions_newVersionAvailable_shouldReturnResult() {
-        testee = new VersionComparator(List.of(getResponse_apache_commons_lang3_v3120()), false);
+        when(settings.isMajorVersionChangeIgnored()).thenReturn(false);
+        testee = new VersionComparator(List.of(getResponse_apache_commons_lang3_v3120()), settings);
         List<DependencyUpdateResult> result = testee.compareDependencyVersions(List.of(createModuleDependency("3.11.0")), Collections.emptyList());
         assertFalse(result.isEmpty());
         assertEquals("org.apache.commons", result.getFirst().getGroupId());
@@ -33,21 +47,22 @@ class VersionComparatorTest {
 
     @Test
     void compareDependencyVersions_noVersionFound_shouldReturnEmptyResult() {
-        testee = new VersionComparator(Collections.emptyList(), false);
+        testee = new VersionComparator(Collections.emptyList(), settings);
         List<DependencyUpdateResult> result = testee.compareDependencyVersions(List.of(createModuleDependency("3.11.0")), Collections.emptyList());
         assertTrue(result.isEmpty());
     }
 
     @Test
     void compareDependencyVersions_alreadyLatestVersionUsed_shouldReturnEmptyResult() {
-        testee = new VersionComparator(Collections.emptyList(), false);
+        testee = new VersionComparator(Collections.emptyList(), settings);
         List<DependencyUpdateResult> result = testee.compareDependencyVersions(List.of(createModuleDependency("3.12.0")), Collections.emptyList());
         assertTrue(result.isEmpty());
     }
 
     @Test
     void compareDependencyVersions_wrongVersionUsed_shouldReturnEmptyResult() {
-        testee = new VersionComparator(List.of(getResponse_apache_commons_lang3_v3120()), false);
+        when(settings.isMajorVersionChangeIgnored()).thenReturn(false);
+        testee = new VersionComparator(List.of(getResponse_apache_commons_lang3_v3120()), settings);
         List<DependencyUpdateResult> result = testee.compareDependencyVersions(List.of(createModuleDependency("3.13.0")), Collections.emptyList());
         assertTrue(result.isEmpty());
     }
@@ -59,7 +74,8 @@ class VersionComparatorTest {
         dependency.setArtifactId("commons-lang3");
         dependency.setVersion("3.11.0");
 
-        testee = new VersionComparator(List.of(getResponse_apache_commons_lang3_v3120()), false);
+        when(settings.isMajorVersionChangeIgnored()).thenReturn(false);
+        testee = new VersionComparator(List.of(getResponse_apache_commons_lang3_v3120()), settings);
         List<DependencyUpdateResult> result = testee.compareDependencyVersions(List.of(createModuleDependency(null)), List.of(dependency));
         assertFalse(result.isEmpty());
     }
@@ -71,7 +87,7 @@ class VersionComparatorTest {
         dependency.setArtifactId("commons-lang3");
         dependency.setVersion(null);
 
-        testee = new VersionComparator(List.of(getResponse_apache_commons_lang3_v3120()), false);
+        testee = new VersionComparator(List.of(getResponse_apache_commons_lang3_v3120()), settings);
         List<DependencyUpdateResult> result = testee.compareDependencyVersions(List.of(createModuleDependency(null)), List.of(dependency));
         assertTrue(result.isEmpty());
     }
@@ -83,21 +99,23 @@ class VersionComparatorTest {
         dependency.setArtifactId("commons-collections4");
         dependency.setVersion("4.4");
 
-        testee = new VersionComparator(List.of(getResponse_apache_commons_lang3_v3120()), false);
+        testee = new VersionComparator(List.of(getResponse_apache_commons_lang3_v3120()), settings);
         List<DependencyUpdateResult> result = testee.compareDependencyVersions(List.of(createModuleDependency(null)), List.of(dependency));
         assertTrue(result.isEmpty());
     }
 
     @Test
     void compareDependencyVersions_checkMinorAndPatchVersionsOnlyEnabled_shouldNotReturnResult() {
-        testee = new VersionComparator(List.of(getResponse_apache_commons_lang3_v500()), true);
+        when(settings.isMajorVersionChangeIgnored()).thenReturn(true);
+        testee = new VersionComparator(List.of(getResponse_apache_commons_lang3_v500()), settings);
         List<DependencyUpdateResult> result = testee.compareDependencyVersions(List.of(createModuleDependency("3.11.0")), Collections.emptyList());
         assertTrue(result.isEmpty());
     }
 
     @Test
     void compareDependencyVersions_checkMinorAndPatchVersionsOnlyEnabledAndMinorVersionAvailable_shouldReturnResult() {
-        testee = new VersionComparator(List.of(getResponse_apache_commons_lang3_v3120(), getResponse_apache_commons_lang3_v500()), true);
+        when(settings.isMajorVersionChangeIgnored()).thenReturn(true);
+        testee = new VersionComparator(List.of(getResponse_apache_commons_lang3_v3120(), getResponse_apache_commons_lang3_v500()), settings);
         List<DependencyUpdateResult> result = testee.compareDependencyVersions(List.of(createModuleDependency("3.11.0")), Collections.emptyList());
         assertFalse(result.isEmpty());
         assertEquals("org.apache.commons", result.getFirst().getGroupId());
@@ -107,8 +125,42 @@ class VersionComparatorTest {
     }
 
     @Test
+    void compareDependencyVersions_checkSnapshotVersion_shouldReturnResult() {
+        when(settings.isPrereleaseVersionsExcluded()).thenReturn(true);
+        testee = new VersionComparator(List.of(getResponse_apache_commons_lang3_v3120_snapshot()), settings);
+        List<DependencyUpdateResult> result = testee.compareDependencyVersions(List.of(createModuleDependency("3.11.0")), Collections.emptyList());
+        assertTrue(result.isEmpty());
+    }
+
+    @Test
+    void compareDependencyVersions_checkDoubleDashVersion_shouldReturnResult() {
+        when(settings.isPrereleaseVersionsExcluded()).thenReturn(true);
+        testee = new VersionComparator(List.of(getResponse_apache_commons_lang3_v3120_doubledash()), settings);
+        List<DependencyUpdateResult> result = testee.compareDependencyVersions(List.of(createModuleDependency("3.11.0")), Collections.emptyList());
+        assertTrue(result.isEmpty());
+    }
+
+    @Test
+    void compareDependencyVersions_checkTimestampVersion_shouldReturnResult() {
+        when(settings.isPrereleaseVersionsExcluded()).thenReturn(true);
+        testee = new VersionComparator(List.of(getResponse_apache_commons_lang3_v3120_timestamp()), settings);
+        List<DependencyUpdateResult> result = testee.compareDependencyVersions(List.of(createModuleDependency("3.11.0")), Collections.emptyList());
+        assertTrue(result.isEmpty());
+    }
+
+    @Test
+    void compareDependencyVersions_checkLetterVersion_shouldReturnResult() {
+        when(settings.isPrereleaseVersionsExcluded()).thenReturn(true);
+        testee = new VersionComparator(List.of(getResponse_apache_commons_lang3_v3120_letter()), settings);
+        List<DependencyUpdateResult> result = testee.compareDependencyVersions(List.of(createModuleDependency("3.11.0")), Collections.emptyList());
+        assertTrue(result.isEmpty());
+    }
+
+    // Plugins //
+
+    @Test
     void comparePluginVersions_newVersionAvailable_shouldReturnResult() {
-        testee = new VersionComparator(List.of(getResponse_maven_compiler_plugin_v3110()), false);
+        testee = new VersionComparator(List.of(getResponse_maven_compiler_plugin_v3110()), settings);
         List<DependencyUpdateResult> result = testee.comparePluginVersions(List.of(createMavenPlugin("3.10.0")), Collections.emptyList());
         assertFalse(result.isEmpty());
         assertEquals("org.apache.maven.plugins", result.getFirst().getGroupId());
@@ -119,14 +171,14 @@ class VersionComparatorTest {
 
     @Test
     void comparePluginVersions_noVersionFound_shouldReturnEmptyResult() {
-        testee = new VersionComparator(Collections.emptyList(), false);
+        testee = new VersionComparator(Collections.emptyList(), settings);
         List<DependencyUpdateResult> result = testee.comparePluginVersions(List.of(createMavenPlugin("3.11.0")), Collections.emptyList());
         assertTrue(result.isEmpty());
     }
 
     @Test
     void comparePluginVersions_alreadyLatestVersionUsed_shouldReturnEmptyResult() {
-        testee = new VersionComparator(Collections.emptyList(), false);
+        testee = new VersionComparator(Collections.emptyList(), settings);
         List<DependencyUpdateResult> result = testee.comparePluginVersions(List.of(createMavenPlugin("3.11.0")), Collections.emptyList());
         assertTrue(result.isEmpty());
     }
@@ -138,13 +190,29 @@ class VersionComparatorTest {
         plugin.setArtifactId("maven-compiler-plugin");
         plugin.setVersion("3.10.0");
 
-        testee = new VersionComparator(List.of(getResponse_maven_compiler_plugin_v3110()), false);
+        testee = new VersionComparator(List.of(getResponse_maven_compiler_plugin_v3110()), settings);
         List<DependencyUpdateResult> result = testee.comparePluginVersions(List.of(createMavenPlugin(null)), List.of(plugin));
         assertFalse(result.isEmpty());
     }
 
     private String getResponse_apache_commons_lang3_v3120() {
         return createJson("org.apache.commons", "commons-lang3", "3.12.0");
+    }
+
+    private String getResponse_apache_commons_lang3_v3120_snapshot() {
+        return createJson("org.apache.commons", "commons-lang3", "3.12.0-SNAPSHOT");
+    }
+
+    private String getResponse_apache_commons_lang3_v3120_timestamp() {
+        return createJson("org.apache.commons", "commons-lang3", "3.12.0+20250719144700");
+    }
+
+    private String getResponse_apache_commons_lang3_v3120_letter() {
+        return createJson("org.apache.commons", "commons-lang3", "3.12.0beta");
+    }
+
+    private String getResponse_apache_commons_lang3_v3120_doubledash() {
+        return createJson("org.apache.commons", "commons-lang3", "3.12.0-beta-2");
     }
 
     private String getResponse_apache_commons_lang3_v500() {